requirements.txt
google-genai
plotly
streamlit
telethon